requirements.txt
Flask==3.0.0
flask-cors==4.0.0

# Para Windows
pywin32==306
Pillow==10.1.0

# Para Linux/Mac
python-escpos==3.0
